major refactor; renames and if guards

+++ compress.py
@@ -1,6 +1,5 @@
 from pathlib import Path
 from subprocess import run
-import subprocess
 import argparse
 from shutil import copyfile
 from typing import Callable
@@ -8,16 +7,16 @@ from typing import Callable
 
 def get_index(index_path: Path) -> int:
     try:
-        with open(index_path, "r") as f:
+        with open(index_path, "r", encoding="utf-8") as f:
             return int(f.read())
-    except:
-        with open(index_path, "w") as f:
+    except FileNotFoundError:
+        with open(index_path, "w", encoding="utf-8") as f:
             f.write("0")
             return 0
 
 
 def update_index(index_path: Path, val: int):
-    with open(index_path, "w") as f:
+    with open(index_path, "w", encoding="utf-8") as f:
         f.write(str(val))
 
 
@@ -26,8 +25,10 @@ def compress_command(in_file: str, out_file: str,  crf: int) -> str:
 
 
 def compress(file: Path, save_file_path: Path, crf: int):
-    completed_process = subprocess.run(compress_command(
-        file, save_file_path, crf))
+    completed_process = run(
+        compress_command(str(file), str(save_file_path), crf),
+        check=False
+    )
 
     if completed_process.returncode:
         raise Exception(f"{file} failed to compress")
@@ -48,39 +49,42 @@ parser.add_argument("-i", "--include", nargs="*", default=[], type=str,
                     help="files to copy into the compressed directory")
 parsed_args = parser.parse_args()
 
-base_path = parsed_args.dir
+base_dir = parsed_args.dir
 crf = parsed_args.crf
 other_files = parsed_args.include
 
 ###################################################
 
-if not base_path.exists():
-    raise FileNotFoundError(f"{base_path} not found")
+if not base_dir.exists():
+    raise FileNotFoundError(f"{base_dir} not found")
 
-save_path = base_path.parent.joinpath(base_path.name + "_compressed")
-save_path.mkdir(exist_ok=True)
+save_dir = base_dir.parent.joinpath(base_dir.name + "_compressed")
+save_dir.mkdir(exist_ok=True)
 
-get_save_file_path = _get_save_file_function(base_path, save_path)
+get_save_file_path = _get_save_file_function(base_dir, save_dir)
 
-files = base_path.rglob("*")
+files = base_dir.rglob("*")
 
 for file in files:
-    dir = get_save_file_path(file).parent
+    directory = get_save_file_path(file).parent
 
-    if not dir.exists():
-        dir.mkdir(parents=True)
+    if not directory.exists():
+        directory.mkdir(parents=True)
 
-index_file_path = base_path.parent.joinpath("index")
+index_file_path = base_dir.parent.joinpath("index")
 index = get_index(index_file_path)
 
-glob = list(base_path.rglob("*.mp4"))
+glob = list(base_dir.rglob("*.mp4"))
 glob_len = len(glob)
 
 
 for file in glob[index:]:
     print(f"{index+1} of {glob_len}")
 
-    compress(file, get_save_file_path(file), crf)
+    save_file = get_save_file_path(file)
+
+    if not save_file.exists():
+        compress(file, save_file, crf)
 
     index += 1
     update_index(index_file_path, index)
@@ -88,9 +92,11 @@ for file in glob[index:]:
 index_file_path.unlink()
 
 for extension in other_files:
-    files = base_path.rglob(extension)
+    files = base_dir.rglob(extension)
 
     for file in files:
-        print(f"Copying {file}")
+        save_file = get_save_file_path(file)
 
-        copyfile(file, get_save_file_path(file))
+        if not save_file.exists():
+            print(f"Copying {file}")
+            copyfile(file, get_save_file_path(file))
